Fixed duplicate dependency issue with Jackson
A big advantage of gradle is the dependency management.

Removing the owltools-oort-all.jar file from the libs directory and
adding the used packages as dependencies in the build.gradle file.

However I commented them out as the packages were not being used. They
only appeared in an unused import statement. As John is working with the
ontology stuff, I didn't completely remove the dependency lines from the
build.gradle so they can be easily added if he needs them.